fixed little gremlin due to i64 handle change (implicit conversion in overloaded function)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@163 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/res/font.cpp
@@ -173,7 +173,7 @@ static int Font_reload(Font* f, const char* fn)
 	// load glyph texture
 	const Handle ht = tex_load(tex_filename);
 	if(ht <= 0)
-		return ht;
+		return (int)ht;
 	tex_upload(ht);
 
 	const int tex_dim = 256;
